RewriteCtx: move removing empty line from Braces
